requirements.txt
streamlit
pandas
numpy
scikit-learn
openpyxl
joblib
pillow